# src/CSharp/EasyMicroservices.StorageMicroservice.Clients/Connected Services/StorageGeneratedServices/OpenAPI.nswag.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EasyMicroservices.StorageMicroservice.WebApi",
    "version": "1.0"
  },
  "paths": {
    "/api/File/UploadFile": {
      "post": {
        "tags": [
          "File"
        ],
        "operationId": "UploadFile",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "File"
                ],
                "properties": {
                  "File": {
                    "type": "string",
                    "format": "binary"
                  },
                  "Key": {
                    "type": "string"
                  },
                  "FolderId": {
                    "title": "Nullable<Int64>",
                    "type": "integer",
                    "format": "int64"
                  },
                  "UniqueIdentity": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "File": {
                  "style": "form"
                },
                "Key": {
                  "style": "form"
                },
                "FolderId": {
                  "style": "form"
                },
                "UniqueIdentity": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FileContractMessageContract"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileContractMessageContract"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileContractMessageContract"
                }
              }
            }
          }
        }
      }
    },
    "/api/File/GetByKeyAndUniqueIdentity": {
      "post": {
        "tags": [
          "File"
        ],
        "operationId": "GetByKeyAndUniqueIdentity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetByKeyRequestContract"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetByKeyRequestContract"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetByKeyRequestContract"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FileContractMessageContract"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileContractMessageContract"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileContractMessageContract"
                }
              }
            }
          }
        }
      }
    },
    "/api/File/GetAllByKeyAndUniqueIdentity": {
      "post": {
        "tags": [
          "File"
        ],
        "operationId": "GetAllByKeyAndUniqueIdentity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetByKeyRequestContract"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetByKeyRequestContract"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetByKeyRequestContract"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FileContractListMessageContract"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileContractListMessageContract"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileContractListMessageContract"
                }
              }
            }
          }
        }
      }
    },
    "/api/File/UploadOrReplaceFile": {
      "post": {
        "tags": [
          "File"
        ],
        "operationId": "UploadOrReplaceFile",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "File"
                ],
                "properties": {
                  "File": {
                    "type": "string",
                    "format": "binary"
                  },
                  "Key": {
                    "type": "string"
                  },
                  "FolderId": {
                    "title": "Nullable<Int64>",
                    "type": "integer",
                    "format": "int64"
                  },
                  "UniqueIdentity": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "File": {
                  "style": "form"
                },
                "Key": {
                  "style": "form"
                },
                "FolderId": {
                  "style": "form"
                },
                "UniqueIdentity": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FileContractMessageContract"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileContractMessageContract"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileContractMessageContract"
                }
              }
            }
          }
        }
      }
    },
    "/api/File/DeleteFileByPassword": {
      "delete": {
        "tags": [
          "File"
        ],
        "operationId": "DeleteFileByPassword",
        "parameters": [
          {
            "name": "fileId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "password",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MessageContract"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageContract"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageContract"
                }
              }
            }
          }
        }
      }
    },
    "/api/File/DownloadFile": {
      "get": {
        "tags": [
          "File"
        ],
        "operationId": "DownloadFile",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "password",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/File/SoftDeleteById": {
      "delete": {
        "tags": [
          "File"
        ],
        "operationId": "SoftDeleteById",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Int64SoftDeleteRequestContract"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Int64SoftDeleteRequestContract"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Int64SoftDeleteRequestContract"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MessageContract"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageContract"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageContract"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/File/GetByUniqueIdentity": {
      "post": {
        "tags": [
          "File"
        ],
        "operationId": "GetByUniqueIdentity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUniqueIdentityRequestContract"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUniqueIdentityRequestContract"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetUniqueIdentityRequestContract"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FileContractMessageContract"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileContractMessageContract"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileContractMessageContract"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/File/GetAllByUniqueIdentity": {
      "post": {
        "tags": [
          "File"
        ],
        "operationId": "GetAllByUniqueIdentity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUniqueIdentityRequestContract"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUniqueIdentityRequestContract"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetUniqueIdentityRequestContract"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FileContractListMessageContract"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileContractListMessageContract"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileContractListMessageContract"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorContract": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "validations": {
            "title": "List<ValidationContract>",
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ValidationContract"
            }
          },
          "failedReasonType": {
            "$ref": "#/components/schemas/FailedReasonType"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "endUserMessage": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "stackTrace": {
            "title": "List<String>",
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "children": {
            "title": "List<ErrorContract>",
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ErrorContract"
            }
          },
          "serviceDetails": {
            "$ref": "#/components/schemas/ServiceDetailsContract"
          }
        }
      },
      "FailedReasonType": {
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "None",
          "Default",
          "All",
          "Other",
          "Unknown",
          "Nothing",
          "SessionAccessDenied",
          "AccessDenied",
          "InternalError",
          "Duplicate",
          "Empty",
          "NotFound",
          "ValidationsError",
          "StreamError",
          "WebServiceNotWorking",
          "Incorrect",
          "OperationFailed"
        ],
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10,
          11,
          12,
          13,
          14,
          15,
          16
        ]
      },
      "FileContract": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "creationDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "key": {
            "type": "string",
            "nullable": true
          },
          "extension": {
            "type": "string",
            "nullable": true
          },
          "length": {
            "type": "integer",
            "format": "int64"
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "contentType": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "folderId": {
            "type": "integer",
            "format": "int64"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "deletedDateTime": {
            "title": "Nullable<DateTime>",
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "downloadLink": {
            "type": "string",
            "nullable": true
          },
          "uniqueIdentity": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "FileContractListMessageContract": {
        "title": "ListMessageContract<FileContract>",
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorContract"
          },
          "success": {
            "$ref": "#/components/schemas/SuccessContract"
          },
          "result": {
            "title": "List<FileContract>",
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/FileContract"
            }
          },
          "hasItems": {
            "type": "boolean",
            "readOnly": true
          }
        }
      },
      "FileContractMessageContract": {
        "title": "MessageContract<FileContract>",
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorContract"
          },
          "success": {
            "$ref": "#/components/schemas/SuccessContract"
          },
          "result": {
            "$ref": "#/components/schemas/FileContract"
          }
        }
      },
...
      "GetByKeyRequestContract": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "uniqueIdentity": {
            "type": "string",
            "nullable": true
          },
          "key": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "GetUniqueIdentityRequestContract": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "uniqueIdentity": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Int64SoftDeleteRequestContract": {
        "title": "SoftDeleteRequestContract<Int64>",
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "isDelete": {
            "type": "boolean"
          }
        }
      },
      "MessageContract": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorContract"
          },
          "success": {
            "$ref": "#/components/schemas/SuccessContract"
          }
        }
      },
      "ServiceDetailsContract": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "servieRouteAddress": {
            "type": "string",
            "nullable": true
          },
          "methodName": {
            "type": "string",
            "nullable": true
          },
          "porjectName": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "SuccessContract": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "endUserMessage": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ValidationContract": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      }
    }
  }
}